requirements.txt
# Core dependencies for n8n video generation setup
flask==2.3.3
requests==2.31.0
pillow==10.0.1
numpy==1.24.3

# Text-to-Speech
gtts==2.4.0
pyttsx3==2.90

# Video processing
opencv-python==4.8.1.78
moviepy==1.0.3

# AI/ML dependencies
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.31.0
accelerate==0.20.3

# Image processing
Pillow==10.0.1

# Audio processing
pydub==0.25.1

# HTTP and networking
urllib3==2.0.4
httpx==0.24.1

# Development and utilities
python-dotenv==1.0.0 